feat(enemies): new basic enemies

+++ script/Entity.cs
@@ -3,9 +3,9 @@ using System;
 
 public partial class Entity : CharacterBody2D
 {
-	[Export] public int Health;
+	[Export] public float Health;
 	
-	public virtual void TakeDamage(int damage) {
+	public virtual void TakeDamage(float damage) {
 		Health = Math.Max(Health - damage, 0);
 		
 		if(Health <= 0) {
